added soap-sending (and signing) jo local tickets (dp_)

+++ ExSoapConnector/SoapAPIConnector/Program.cs
@@ -314,23 +314,42 @@ namespace SoapAPIConnector
                             string thumbPrint = docSettings.Thumpprint != null ? docSettings.Thumpprint : conf.Thumpprint;
                             string body = Utils.Base64Encode(File.ReadAllBytes(name), "windows-1251");
                             string sign = controller.Sign(thumbPrint, body);
-                            if (controller.sendDocApi(body, sign, docType))
-                            {
-                                Logger.log(Path.GetFileName(name) + " sent successfully.");
-                                if (conf.Outbound.IsArchive)
+                            if (docType.StartsWith("DP_"))
+                            {         
+                                if ((controller.sendDoc(Path.GetFileName(name), body)) 
+                                    &&
+                                    (controller.sendDoc(Path.GetFileName(name).Replace(".xml",".bin"), sign)))
                                 {
-                                    if (moveDocToArc(Path.GetFileName(name), File.ReadAllBytes(name)))
-                                        File.Delete(name);
-                                    if (moveDocToArc(Path.GetFileName(name).Replace(".xml", ".bin"), Utils.StringToBytes(sign, "UTF-8")))
-                                        File.Delete(name.Replace(".xml", ".bin"));
+                                    Logger.log(Path.GetFileName(name) + " sent successfully.");
+                                    if (conf.Outbound.IsArchive)
+                                    {
+                                        if ((moveDocToArc(Path.GetFileName(name), (File.ReadAllBytes(name))))
+                                            &&
+                                            (moveDocToArc(Path.GetFileName(name).Replace(".xml", ".bin"), Utils.StringToBytes(sign, "UTF-8"))))
+                                            File.Delete(name);
+                                    }
                                 }
                             }
                             else
                             {
-                                if (moveDocToError(Path.GetFileName(name), File.ReadAllBytes(name)))
-                                    File.Delete(name);
-                                if (moveDocToError(Path.GetFileName(name).Replace(".xml", ".bin"), Utils.StringToBytes(sign, "UTF-8")))
-                                    File.Delete(name.Replace(".xml", ".bin"));
+                                if (controller.sendDocApi(body, sign, docType))
+                                {
+                                    Logger.log(Path.GetFileName(name) + " sent successfully.");
+                                    if (conf.Outbound.IsArchive)
+                                    {
+                                        if (moveDocToArc(Path.GetFileName(name), File.ReadAllBytes(name)))
+                                            File.Delete(name);
+                                        if (moveDocToArc(Path.GetFileName(name).Replace(".xml", ".bin"), Utils.StringToBytes(sign, "UTF-8")))
+                                            File.Delete(name.Replace(".xml", ".bin"));
+                                    }
+                                }
+                                else
+                                {
+                                    if (moveDocToError(Path.GetFileName(name), File.ReadAllBytes(name)))
+                                        File.Delete(name);
+                                    if (moveDocToError(Path.GetFileName(name).Replace(".xml", ".bin"), Utils.StringToBytes(sign, "UTF-8")))
+                                        File.Delete(name.Replace(".xml", ".bin"));
+                                }
                             }
 
                         }
